# api/docs/swagger.yaml
definitions:
  models.Book:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateBook:
    properties:
      created_at:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  models.CreateCategory:
    properties:
      created_at:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Empty:
    type: object
  models.GetListBookResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/models.Book'
        type: array
      count:
        type: integer
    type: object
  models.GetListCategoryResponse:
    properties:
      categorys:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      count:
        type: integer
    type: object
  models.UpdateBookId:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  models.UpdateCategoryId:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /book:
    get:
      consumes:
      - application/json
      description: Get List Book
      operationId: BookPrimeryKey
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetBookListBody
          schema:
            $ref: '#/definitions/models.GetListBookResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Book
      tags:
      - Book
    post:
      consumes:
      - application/json
      description: CreateBook
      operationId: CreateBook
      parameters:
      - description: CreateBookRequestBody
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.CreateBook'
      produces:
      - application/json
      responses:
        "201":
          description: GetBookBody
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateBook
      tags:
      - Book
  /book/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Book
      operationId: DeleteBook
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteBookBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Book
      tags:
      - Book
    get:
      consumes:
      - application/json
      description: GetByID Book
      operationId: Get_By_IDBook
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: GetByIDBookBody
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: GetByID Book
      tags:
      - Book
    put:
      consumes:
      - application/json
      description: Update Book
      operationId: UpdateBook
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateBookRequestBody
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBookId'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateBookBody
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Book
      tags:
      - Book
  /category:
    get:
      consumes:
      - application/json
      description: Get List Category
      operationId: get_list_category
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetCategoryListBody
          schema:
            $ref: '#/definitions/models.GetListCategoryResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Category
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create Category
      operationId: create_category
      parameters:
      - description: CreateCategoryRequestBody
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CreateCategory'
      produces:
      - application/json
      responses:
        "201":
          description: GetCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Category
      tags:
      - Category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Category
      operationId: delete_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteCategoryBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get By ID Category
      operationId: get_by_id_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update Category
      operationId: update_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCategoryRequestBody
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryId'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Category
      tags:
      - Category
swagger: "2.0"